requirements.txt
matplotlib
numpy
pytorch 1.10
scipy